add map stories marker

+++ app/src/main/java/com/dwi/dicodingstoryapp/data/source/remote/response/StoriesResponse.kt
@@ -33,11 +33,11 @@ data class StoryResult(
 	val description: String? = null,
 
 	@field:SerializedName("lon")
-	val lon: Float? = null,
+	val lon: Double? = null,
 
 	@field:SerializedName("id")
 	val id: String? = null,
 
 	@field:SerializedName("lat")
-	val lat: Float? = null
+	val lat: Double? = null
 ) : Parcelable
